Convert link check extension to parameterized source
